BackEnd/requirements.txt
django>=5.0.0
djangorestframework
django-cors-headers
Pillow
